Revert "Split the SceneMapIdSceneAsset to its own file. (#5400)"
This reverts commit 3c4aa27d5a981ccdab2220a2ff50b4d49e14d0ad.
