--- com.unity.render-pipelines.high-definition/Runtime/Utilities/SceneObjectIDMapSceneAsset.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0c501ba6f44a1a44ca76ed3c3be275cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
